Move identity pallet custom types into its own module
